requirements.txt
# requirements.txt

connexion[Flask]==3.1.0
connexion[uvicorn]==3.1.0
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
marshmallow-sqlalchemy==1.1.0
swagger_ui_bundle==1.1.0
